views.py
views.py is another important file. Views are the 'V' of MVT. Views provide an interface through which users interact with a Django website. It connects models and templates together.
In this file, you write the business logic for your app. A view can be either function-based or class-based. You decide if you want to write your views using functions or classes.